refactor: u8 to u32 for entries_left/entry_limit, rename tournament to context
- Change entries_left return type from Option<u8> to Option<u32>
- Change add_config entry_limit param from u8 to u32
- Update all storage maps, variables, and arithmetic accordingly
- Rename tournament_ prefixed storage/variables to context_ in all
  non-tournament validators (snapshot, erc20, opus, governance, zk, merkle)
- Update test mocks and test files to match

+++ packages/entry_requirement/src/entry_requirement_extension_component.cairo
@@ -54,11 +54,11 @@ pub mod EntryRequirementExtensionComponent {
             context_id: u64,
             player_address: ContractAddress,
             qualification: Span<felt252>,
-        ) -> Option<u8>;
+        ) -> Option<u32>;
 
         /// Add configuration for a context
         fn add_config(
-            ref self: TContractState, context_id: u64, entry_limit: u8, config: Span<felt252>,
+            ref self: TContractState, context_id: u64, entry_limit: u32, config: Span<felt252>,
         );
 
         /// Called when an entry is added
@@ -128,7 +128,7 @@ pub mod EntryRequirementExtensionComponent {
             context_id: u64,
             player_address: ContractAddress,
             qualification: Span<felt252>,
-        ) -> Option<u8> {
+        ) -> Option<u32> {
             let contract = self.get_contract();
             EntryRequirementExtension::entries_left(
                 contract, context_id, player_address, qualification,
@@ -138,7 +138,7 @@ pub mod EntryRequirementExtensionComponent {
         fn add_config(
             ref self: ComponentState<TContractState>,
             context_id: u64,
-            entry_limit: u8,
+            entry_limit: u32,
             config: Span<felt252>,
         ) {
             self.set_context_owner(context_id);
